feat(credoffer): added ui for new cred offer invalid credentials
