--- src/ui/home/containers/credentialReceive.tsx
@@ -1,197 +1,0 @@
-import React, { useState } from 'react'
-import { connect } from 'react-redux'
-import {
-  Image,
-  ScrollView,
-  StyleSheet,
-  Text,
-  TouchableOpacity,
-  View,
-} from 'react-native'
-import { ThunkDispatch } from '../../../store'
-import {
-  acceptSelectedCredentials,
-  CredentialOfferRenderDetails,
-} from '../../../actions/sso/credentialOfferRequest'
-import { CredentialOfferResponseSelection } from 'jolocom-lib/js/interactionTokens/interactionTokens.types'
-import { NavigationScreenProp, NavigationState } from 'react-navigation'
-import { CredentialOfferRequest } from 'jolocom-lib/js/interactionTokens/credentialOfferRequest'
-import { IdentitySummary } from '../../../actions/sso/types'
-import { JSONWebToken } from 'jolocom-lib/js/interactionTokens/JSONWebToken'
-import { withErrorScreen, withLoading } from '../../../actions/modifiers'
-import { navigationActions } from '../../../actions'
-import { routeList } from '../../../routeList'
-import { DocumentCard } from '../../documents/components/documentCard'
-import { JolocomButton, Wrapper } from '../../structure'
-import { Colors, Spacing } from '../../../styles'
-import { centeredText, fontMain, fontMedium } from '../../../styles/typography'
-import { black065, overflowBlack } from '../../../styles/colors'
-import { ActionSheet } from '../../structure/actionSheet'
-import strings from '../../../locales/strings'
-import I18n from 'src/locales/i18n'
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: Colors.lightGreyLighter,
-  },
-  topSection: {
-    paddingVertical: Spacing.MD,
-    alignItems: 'center',
-  },
-  logo: {
-    borderRadius: 35,
-    width: 70,
-    height: 70,
-    margin: 10,
-  },
-  serviceName: {
-    fontFamily: fontMedium,
-    fontSize: 28,
-    color: overflowBlack,
-  },
-  description: {
-    fontFamily: fontMain,
-    fontSize: 16,
-    color: black065,
-    marginTop: 10,
-    marginBottom: 4,
-    marginHorizontal: '10%',
-    ...centeredText,
-  },
-  documentWrapper: {
-    alignItems: 'center',
-    padding: 10,
-  },
-})
-
-export interface CredentialOfferNavigationParams {
-  credentialOfferRequest: JSONWebToken<CredentialOfferRequest>
-  requesterSummary: IdentitySummary
-  credentialRenderDetails: CredentialOfferRenderDetails[]
-  isDeepLink: boolean
-}
-
-interface Props extends ReturnType<typeof mapDispatchToProps> {
-  navigation: NavigationScreenProp<
-    NavigationState,
-    CredentialOfferNavigationParams
-  >
-}
-
-export const CredentialsReceiveContainer = (props: Props) => {
-  const [selected, setSelected] = useState<CredentialOfferRenderDetails[]>([])
-  const { navigation, acceptSelectedCredentials, goBack } = props
-  const {
-    state: {
-      params: {
-        credentialOfferRequest,
-        requesterSummary,
-        isDeepLink,
-        credentialRenderDetails,
-      },
-    },
-  } = navigation
-  const publicProfile = requesterSummary && requesterSummary.publicProfile
-
-  const handleConfirm = () => {
-    const responseSelection: CredentialOfferResponseSelection[] = selected.map(
-      credential => ({
-        type: credential.type,
-      }),
-    )
-    acceptSelectedCredentials(
-      responseSelection,
-      credentialOfferRequest,
-      isDeepLink,
-    )
-  }
-
-  const onPressDocument = (cred: CredentialOfferRenderDetails) => {
-    if (selected.includes(cred)) {
-      setSelected(selected.filter(current => current !== cred))
-    } else {
-      setSelected([...selected, cred])
-    }
-  }
-
-  return (
-    <Wrapper style={{ backgroundColor: Colors.iBackgroundWhite }}>
-      <ScrollView
-        contentContainerStyle={{ paddingBottom: '50%' }}
-        style={{ width: '100%' }}
-      >
-        {publicProfile && (
-          <View style={styles.topSection}>
-            <Image style={styles.logo} source={{ uri: publicProfile.image }} />
-            <Text style={styles.serviceName}>{publicProfile.name}</Text>
-            <Text style={styles.description}>
-              {I18n.t(
-                strings.CHOOSE_ONE_OR_MORE_DOCUMENTS_PROVIDED_BY_THIS_SERVICE_AND_WE_WILL_GENERATE_THEM_FOR_YOU,
-              )}
-            </Text>
-          </View>
-        )}
-        {credentialRenderDetails.map(summary => {
-          const { type, renderInfo } = summary
-          const isSelected = selected.includes(summary)
-          return (
-            <TouchableOpacity
-              onPress={() => onPressDocument(summary)}
-              activeOpacity={1}
-              style={styles.documentWrapper}
-            >
-              <DocumentCard
-                selected={isSelected}
-                credentialType={type}
-                renderInfo={renderInfo}
-              />
-            </TouchableOpacity>
-          )
-        })}
-      </ScrollView>
-      <ActionSheet showSlide={true}>
-        <JolocomButton
-          textStyle={{ fontFamily: fontMedium }}
-          disabled={selected.length === 0}
-          onPress={handleConfirm}
-          text={I18n.t(strings.RECEIVE)}
-        />
-        <JolocomButton
-          containerStyle={{ marginTop: 10 }}
-          onPress={goBack}
-          text={I18n.t(strings.CANCEL)}
-          transparent
-        />
-      </ActionSheet>
-    </Wrapper>
-  )
-}
-
-const mapDispatchToProps = (dispatch: ThunkDispatch) => ({
-  acceptSelectedCredentials: (
-    selected: CredentialOfferResponseSelection[],
-    credentialOfferRequest: JSONWebToken<CredentialOfferRequest>,
-    isDeepLink: boolean,
-  ) =>
-    dispatch(
-      withErrorScreen(
-        withLoading(
-          acceptSelectedCredentials(
-            selected,
-            credentialOfferRequest,
-            isDeepLink,
-          ),
-        ),
-      ),
-    ),
-  goBack: () =>
-    dispatch(
-      navigationActions.navigate({ routeName: routeList.InteractionScreen }),
-    ),
-})
-
-export const CredentialReceive = connect(
-  null,
-  mapDispatchToProps,
-)(CredentialsReceiveContainer)
