+++ src/ui/home/index.tsx
@@ -2,4 +2,4 @@ export { Claims } from 'src/ui/home/containers/claims'
 export { ClaimDetails } from 'src/ui/home/containers/claimDetails'
 export { Records } from 'src/ui/home/containers/records'
 export { Consent } from 'src/ui/sso/containers/consent'
-export { CredentialReceive } from 'src/ui/home/containers/credentialReceive'
+export { CredentialReceive } from 'src/ui/sso/containers/credentialReceive'
